@@ -3,23 +3,24 @@
     "license": "MIT",
     "version": "1.8.2",
     "require": {
-        "php": "~8.0.0 | ~8.1.0",
+        "php": "^8.2",
         "ext-ctype": "*",
         "ext-iconv": "*",
         "ext-json": "*",
-        "composer/package-versions-deprecated": "1.11.99.5",
-        "doctrine/doctrine-bundle": "^2",
-        "doctrine/doctrine-migrations-bundle": "^2",
-        "doctrine/orm": "^2",
-        "nelmio/cors-bundle": "^1.5 | ^2.0",
-        "symfony/console": "^5.4",
-        "symfony/dotenv": "^5.4",
-        "symfony/flex": "^1.2",
-        "symfony/framework-bundle": "^5.4",
-        "symfony/yaml": "^5.4"
+        "doctrine/cache": "^2.2",
+        "doctrine/doctrine-bundle": "^2.12",
+        "doctrine/doctrine-migrations-bundle": "^3.3",
+        "doctrine/orm": "^3.0",
+        "nelmio/cors-bundle": "^2.4",
+        "symfony/console": "^7.3",
+        "symfony/dotenv": "^7.3",
+        "symfony/flex": "^2.4",
+        "symfony/framework-bundle": "^7.3",
+        "symfony/yaml": "^7.3"
     },
     "require-dev": {
-        "symfony/maker-bundle": "^1.6"
+            "symfony/maker-bundle": "^1.60",
+            "rector/rector": "^1.0"
     },
     "config": {
         "preferred-install": {
@@ -27,7 +28,8 @@
         },
         "sort-packages": true,
         "allow-plugins": {
-            "symfony/flex": true
+                "symfony/flex": true,
+                "symfony/runtime": true
         }
     },
     "autoload": {
@@ -64,7 +66,8 @@
     },
     "extra": {
         "symfony": {
-            "allow-contrib": false
+            "allow-contrib": false,
+            "require": "7.3.*"
         }
     }
 }
