@@ -12,11 +12,9 @@ use App\Service\TransactionService;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Attribute\Route;
 
-/**
- * @Route("/api")
- */
+#[Route('/api')]
 class TransactionController extends AbstractController {
 
     /**
@@ -28,9 +26,7 @@ class TransactionController extends AbstractController {
         $this->transactionSerializer = $transactionSerializer;
     }
 
-    /**
-     * @Route("/transaction", methods="GET")
-     */
+    #[Route('/transaction', methods: ['GET'])]
     function list(Request $request, EntityManagerInterface $entityManager) {
         $limit = $request->query->get('limit', 25);
         $offset = $request->query->get('offset');
@@ -46,9 +42,7 @@ class TransactionController extends AbstractController {
         ]);
     }
 
-    /**
-     * @Route("/user/{userId}/transaction", methods="POST")
-     */
+    #[Route('/user/{userId}/transaction', methods: ['POST'])]
     function createUserTransactions($userId, Request $request, TransactionService $transactionService, EntityManagerInterface $entityManager) {
         $amount = $request->request->get('amount');
         $quantity = $request->request->get('quantity');
@@ -72,9 +66,7 @@ class TransactionController extends AbstractController {
         ]);
     }
 
-    /**
-     * @Route("/user/{userId}/transaction", methods="GET")
-     */
+    #[Route('/user/{userId}/transaction', methods: ['GET'])]
     function getUserTransactions($userId, Request $request, EntityManagerInterface $entityManager) {
         $limit = $request->query->get('limit', 25);
         $offset = $request->query->get('offset');
@@ -95,9 +87,7 @@ class TransactionController extends AbstractController {
         ]);
     }
 
-    /**
-     * @Route("/user/{userId}/transaction/{transactionId}", methods="GET")
-     */
+    #[Route('/user/{userId}/transaction/{transactionId}', methods: ['GET'])]
     function getUserTransaction($userId, $transactionId, EntityManagerInterface $entityManager) {
         $user = $entityManager->getRepository(User::class)->find($userId);
         if (!$user) {
@@ -114,9 +104,7 @@ class TransactionController extends AbstractController {
         ]);
     }
 
-    /**
-     * @Route("/user/{userId}/transaction/{transactionId}", methods="DELETE")
-     */
+    #[Route('/user/{userId}/transaction/{transactionId}', methods: ['DELETE'])]
     function deleteTransaction($userId, $transactionId, TransactionService $transactionService) {
         $transaction = $transactionService->revertTransaction($transactionId);
 
